Re-organize views.py into several files
